Added features to support user defined metatypes for C structures/unions or complex types.
The idea is to allow for type level attributes which are available in all instances of a type.
Somewhat similar to C++ static members of a class.

One can add a metatble to type using ffi.metatype and then access the metamethods in the instances.

This is inspired from similar features available in OpenResty.
However some of the features supporting C++ syntax are not part.